backend/requirements.txt
# FastAPI 核心
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# 数据库
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.0

# 认证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# Redis
redis>=5.0.0

# Ansible
ansible-runner>=2.3.6
ansible-core>=2.16.0

# 工具库
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
email-validator>=2.1.0

# 定时任务
apscheduler>=3.10.0

# WebSocket
websockets>=12.0

# 加密
cryptography>=41.0.0
aiohttp>=3.9.0
paramiko>=3.4.0
